requirements.txt
torch>=2.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
scikit-surprise>=1.1.3
matplotlib>=3.7
pytest>=7.0
